-add a test for FindBlitz.cmake -change output prefix from ">>" to "::" (doesn't have to be quoted when grepping) -add option CLEAN_DIRS to RunAllModuleTests.cmake, if you set this to TRUE, the previous test build directories will be removed before running the tests
Alex


svn path=/trunk/KDE/kdelibs/; revision=884400

+++ modules-tests/RunAllModuleTests.cmake
@@ -1,13 +1,16 @@
 # run this script via "cmake -P RunAllModuletests.cmake" to
 # execute all module tests in one go.
 # To see only the results and not the other cmake output,
-# you can grep for ">>".
+# you can grep for "::".
+# To have it delete the previous test build dirs, run it with -DCLEAN_DIRS=TRUE
 
 get_filename_component(currentDir "${CMAKE_CURRENT_LIST_FILE}" PATH)
 
 function(execute_one_test name)
    set(workingDir "${currentDir}/${name}/build")
-   file(REMOVE_RECURSE "${workingDir}")
+   if(CLEAN_DIRS)
+      file(REMOVE_RECURSE "${workingDir}")
+   endif(CLEAN_DIRS)
    file(MAKE_DIRECTORY "${workingDir}")
    execute_process(COMMAND ${CMAKE_COMMAND} "${currentDir}/${name}"
                    WORKING_DIRECTORY "${workingDir}")
@@ -16,5 +19,6 @@ endfunction(execute_one_test)
 execute_one_test(Xine)
 execute_one_test(Flex)
 execute_one_test(QCA2)
+execute_one_test(Blitz)
